Rename package 'awapi_accessor' to 'concealer'
