+++ openscenario/openscenario_interpreter/package.xml
@@ -9,7 +9,7 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <depend>awapi_accessor</depend>
+  <depend>concealer</depend>
   <depend>geometry_msgs</depend>
   <depend>junit_exporter</depend>
   <depend>libgoogle-glog-dev</depend>
@@ -19,8 +19,8 @@
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
   <depend>rclcpp_lifecycle</depend>
-  <depend>traffic_simulator</depend>
   <depend>std_msgs</depend>
+  <depend>traffic_simulator</depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_lint_auto</test_depend>
