+++ simulation/traffic_simulator/package.xml
@@ -13,8 +13,8 @@
 
   <depend>ament_index_cpp</depend>
   <depend>autoware_auto_msgs</depend>
-  <depend>awapi_accessor</depend>
   <depend>behaviortree_cpp_v3</depend>
+  <depend>concealer</depend>
   <depend>geographic_msgs</depend>
   <depend>geometry_msgs</depend>
   <depend>lanelet2_core</depend>
@@ -22,18 +22,18 @@
   <depend>lanelet2_io</depend>
   <depend>lanelet2_projection</depend>
   <depend>lanelet2_routing</depend>
+  <depend>nlohmann-json-dev</depend>
   <depend>openscenario_msgs</depend>
   <depend>pugixml-dev</depend>
   <depend>quaternion_operation</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
   <depend>rviz2</depend>
+  <depend>simulation_interface</depend>
   <depend>std_msgs</depend>
   <depend>tf2_geometry_msgs</depend>
   <depend>tf2_ros</depend>
   <depend>visualization_msgs</depend>
-  <depend>simulation_interface</depend>
-  <depend>nlohmann-json-dev</depend>
 
   <!--<exec_depend>lexus_rx_450h_description</exec_depend> -->
   <exec_depend>kashiwanoha_map</exec_depend>
